storage: don't fail on non existent storage entities

+++ lxd/storage_lvm.go
@@ -331,8 +331,10 @@ func (s *storageLvm) StoragePoolDelete() error {
 	}
 
 	poolName := s.getOnDiskPoolName()
+	poolExists, _ := storageVGExists(poolName)
+
 	// Delete the thinpool.
-	if s.useThinpool {
+	if s.useThinpool && poolExists {
 		// Check that the thinpool actually exists. For example, it
 		// won't when the user has never created a storage volume in the
 		// storage pool.
@@ -357,7 +359,7 @@ func (s *storageLvm) StoragePoolDelete() error {
 	}
 
 	// Remove the volume group.
-	if count == 0 {
+	if count == 0 && poolExists {
 		output, err := shared.TryRunCommand("vgremove", "-f", poolName)
 		if err != nil {
 			logger.Errorf("failed to destroy the volume group for the lvm storage pool: %s", output)
@@ -521,10 +523,16 @@ func (s *storageLvm) StoragePoolVolumeCreate() error {
 func (s *storageLvm) StoragePoolVolumeDelete() error {
 	logger.Infof("Deleting LVM storage volume \"%s\" on storage pool \"%s\".", s.volume.Name, s.pool.Name)
 
-	customPoolVolumeMntPoint := getStoragePoolVolumeMountPoint(s.pool.Name, s.volume.Name)
-	_, err := s.StoragePoolVolumeUmount()
-	if err != nil {
-		return err
+	poolName := s.getOnDiskPoolName()
+	customLvmDevPath := getLvmDevPath(poolName,
+		storagePoolVolumeAPIEndpointCustom, s.volume.Name)
+	lvExists, _ := storageLVExists(customLvmDevPath)
+
+	if lvExists {
+		_, err := s.StoragePoolVolumeUmount()
+		if err != nil {
+			return err
+		}
 	}
 
 	volumeType, err := storagePoolVolumeTypeNameToAPIEndpoint(s.volume.Type)
@@ -532,12 +540,14 @@ func (s *storageLvm) StoragePoolVolumeDelete() error {
 		return err
 	}
 
-	poolName := s.getOnDiskPoolName()
-	err = s.removeLV(poolName, volumeType, s.volume.Name)
-	if err != nil {
-		return err
+	if lvExists {
+		err = s.removeLV(poolName, volumeType, s.volume.Name)
+		if err != nil {
+			return err
+		}
 	}
 
+	customPoolVolumeMntPoint := getStoragePoolVolumeMountPoint(s.pool.Name, s.volume.Name)
 	if shared.PathExists(customPoolVolumeMntPoint) {
 		err := os.Remove(customPoolVolumeMntPoint)
 		if err != nil {
@@ -998,21 +1008,27 @@ func (s *storageLvm) ContainerDelete(container container) error {
 	}
 
 	poolName := s.getOnDiskPoolName()
-	err := s.removeLV(poolName, storagePoolVolumeAPIEndpointContainers, containerLvmName)
-	if err != nil {
-		return err
+	containerLvmDevPath := getLvmDevPath(poolName,
+		storagePoolVolumeAPIEndpointContainers, containerLvmName)
+	lvExists, _ := storageLVExists(containerLvmDevPath)
+
+	if lvExists {
+		err := s.removeLV(poolName, storagePoolVolumeAPIEndpointContainers, containerLvmName)
+		if err != nil {
+			return err
+		}
 	}
 
 	if container.IsSnapshot() {
 		sourceName, _, _ := containerGetParentAndSnapshotName(containerName)
 		snapshotMntPointSymlinkTarget := shared.VarPath("storage-pools", s.pool.Name, "snapshots", sourceName)
 		snapshotMntPointSymlink := shared.VarPath("snapshots", sourceName)
-		err = deleteSnapshotMountpoint(containerMntPoint, snapshotMntPointSymlinkTarget, snapshotMntPointSymlink)
+		err := deleteSnapshotMountpoint(containerMntPoint, snapshotMntPointSymlinkTarget, snapshotMntPointSymlink)
 		if err != nil {
 			return err
 		}
 	} else {
-		err = deleteContainerMountpoint(containerMntPoint, container.Path(), s.GetStorageTypeName())
+		err := deleteContainerMountpoint(containerMntPoint, container.Path(), s.GetStorageTypeName())
 		if err != nil {
 			return err
 		}
@@ -1567,15 +1583,21 @@ func (s *storageLvm) ImageDelete(fingerprint string) error {
 	logger.Debugf("Deleting LVM storage volume for image \"%s\" on storage pool \"%s\".", fingerprint, s.pool.Name)
 
 	if s.useThinpool {
-		_, err := s.ImageUmount(fingerprint)
-		if err != nil {
-			return err
-		}
-
 		poolName := s.getOnDiskPoolName()
-		err = s.removeLV(poolName, storagePoolVolumeAPIEndpointImages, fingerprint)
-		if err != nil {
-			return err
+		imageLvmDevPath := getLvmDevPath(poolName,
+			storagePoolVolumeAPIEndpointImages, fingerprint)
+		lvExists, _ := storageLVExists(imageLvmDevPath)
+
+		if lvExists {
+			_, err := s.ImageUmount(fingerprint)
+			if err != nil {
+				return err
+			}
+
+			err = s.removeLV(poolName, storagePoolVolumeAPIEndpointImages, fingerprint)
+			if err != nil {
+				return err
+			}
 		}
 	}
